requirements.txt
python-irclib
